refactor: 게임 및 타이머 로직을 커스텀 훅으로 분리
